Loading + Uploading page
The courses and users files can be uploaded in 'view' page, and a confirmation is required on 'upload' page. Before, the same page has the confirmation button and the really upload functions, but now we have an another page called 'uploading_data' to where the confirmation send the information and where the data is really uploaded. It has a loading view too, but the process of load + upload isn't configured.