Reestruturing.
Now we use '.mustache' templates to write views.

In utils directory we have some functions to help in development, like a less bulky 'render_template', CSV utilities, forms, tables etc.

In 'models' directory we have functions to work with database. Moodle isn't MVC, but for now consider it as if it were.

The pages have been reestructured to comport the templates and some improvments with Moodle functions.